[Mapping.NonLinear] Move DistanceMultiMapping in its own files
