+++ Sofa/Component/Mapping/NonLinear/CMakeLists.txt
@@ -10,6 +10,8 @@ set(HEADER_FILES
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceFromTargetMapping.inl
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMapping.h
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMapping.inl
+    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMultiMapping.h
+    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMultiMapping.inl
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/RigidMapping.h
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/RigidMapping.inl
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/RigidRigidMapping.h
@@ -24,6 +26,7 @@ set(SOURCE_FILES
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/init.cpp
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceFromTargetMapping.cpp
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMapping.cpp
+    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMultiMapping.cpp
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/RigidMapping.cpp
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/RigidRigidMapping.cpp
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/SquareDistanceMapping.cpp
